fix for cue/bin file reads being very slow due to reading CD_SECTOR_SIZE_AUDIO at a time.
fopen on switch buffers only 1k, and CD_SECTOR_SIZE_AUDIO is 2k. so when reading a 256k chunk,
this will cause 128 reads.
this change makes fopen buffer 256k at a time, so in the above example, only 1 read would happen.

a proper fix would be to read directly into the out buffer, and on the last block read into the local
buffer and copy it over (if not aligned to CD_SECTOR_SIZE_AUDIO).
this will fix the above without the need of setvbuf, and will massively improve audio cd throughput.